Combine blending and crop into a module called render
