+++ src/minerva_lib/render.py
@@ -1,7 +1,90 @@
 import itertools
 import numpy as np
-from .blend import composite_channel
 from . import skimage_inline as ski
+
+
+def composite_channel(target, image, color, range_min, range_max, out=None):
+    ''' Render _image_ in pseudocolor and composite into _target_
+
+    By default, a new output array will be allocated to hold
+    the result of the composition operation. To update _target_
+    in place instead, specify the same array for _target_ and _out_.
+
+    Args:
+        target: Numpy array containing composition target image
+        image: Numpy array of image to render and composite
+        color: Color as r, g, b float array within 0, 1
+        range_min: Threshhold range minimum, float within 0, 1
+        range_max: Threshhold range maximum, float within 0, 1
+        out: Optional output numpy array in which to place the result.
+
+    Returns:
+        A numpy array with the same shape as the composited image.
+        If an output array is specified, a reference to _out_ is returned.
+    '''
+
+    if out is None:
+        out = target.copy()
+
+    # Rescale the new channel to a float64 between 0 and 1
+    f64_range = (range_min, range_max)
+    f64_image = ski.img_as_float(image)
+    f64_image = ski.rescale_intensity(f64_image, f64_range)
+
+    # Colorize and add the new channel to composite image
+    for i, component in enumerate(color):
+        out[:, :, i] += f64_image * component
+
+    return out
+
+
+def composite_channels(channels):
+    '''Render each image in _channels_ additively into a composited image
+
+    Args:
+        channels: List of dicts for channels to blend. Each dict in the
+            list must have the following rendering settings:
+            {
+                image: Numpy 2D image data of any type
+                color: Color as r, g, b float array within 0, 1
+                min: Threshhold range minimum, float within 0, 1
+                max: Threshhold range maximum, float within 0, 1
+            }
+
+    Returns:
+        For input images with shape `(n,m)`,
+        returns a float32 RGB color image with shape
+        `(n,m,3)` and values in the range 0 to 1
+    '''
+
+    num_channels = len(channels)
+
+    # Must be at least one channel
+    if num_channels < 1:
+        raise ValueError('At least one channel must be specified')
+
+    # Ensure that dimensions of all channels are equal
+    shape = channels[0]['image'].shape
+    for channel in channels:
+        if channel['image'].shape != shape:
+            raise ValueError('All channel images must have equal dimensions')
+
+    # Shape of 3 color image
+    shape_color = shape + (3,)
+
+    # Final buffer for blending
+    out_buffer = np.zeros(shape_color, dtype=np.float32)
+
+    # rescaled images and normalized colors
+    for channel in channels:
+
+        # Add all three channels to output buffer
+        args = map(channel.get, ['image', 'color', 'min', 'max'])
+        composite_channel(out_buffer, *args, out=out_buffer)
+
+    # Return gamma correct image within 0, 1
+    np.clip(out_buffer, 0, 1, out=out_buffer)
+    return ski.adjust_gamma(out_buffer, 1 / 2.2)
 
 
 def scale_image_nearest_neighbor(source, factor):
